Arreglado bug mergeDebugAndroidTestJavaResource

+++ domain/inventory/src/main/java/app/domain/invoicing/dao/ProductDao.kt
@@ -24,7 +24,7 @@ interface ProductDao {
     fun getAllProducts(): Flow<List<Product>>
 
     @Query("SELECT * FROM Product WHERE id = :productId")
-    fun getProductById(productId: Int): Product?
+    fun getProductById(productId: Long): Product?
 
     @Query("SELECT * FROM Product WHERE code = :productCode")
     fun validate(productCode: String): Product?
